refactor: move deleted helpers to models module
Consolidates DELETED_PREFIX, is_deleted(), mark_deleted(), and
unmark_deleted() in models.py to eliminate duplication between
cli.py and export.py.
